# docs/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: SignVault API
  version: 0.2.0
  description: |
    Kontrak API SignVault v0.2 (sumber kebenaran, prd.md §10).
    Scope: backend (apps/backend, packages/database, docs/api-docs).
    v0.2: Google OAuth (POST /auth/google), onboarding (PATCH
    /users/me/complete-profile), CAPTCHA Turnstile, rate-limit auth 10/mnt.
    Base path: /api/v1.
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev (Uvicorn native, tanpa Docker)
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Registrasi akun baru (6 field)
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '201':
          description: 'Terdaftar, email verifikasi terkirim (mode dev: link ada di log server)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Verifikasi email dengan token sekali pakai (expired 24 jam)
      operationId: verifyEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Email terverifikasi, akun aktif
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/resend-verification:
    post:
      tags: [auth]
      summary: Kirim ulang email verifikasi
      operationId: resendVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Email terkirim (selalu 200 agar tidak bocor status akun)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/login:
    post:
      tags: [auth]
      summary: Login dengan email terverifikasi
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Login sukses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Email belum terverifikasi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Tukar refresh token menjadi pasangan token baru (rotation)
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Token baru
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout (revoke refresh token)
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Logout sukses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/google:
    post:
      tags: [auth]
      summary: Login/daftar via Google ID Token (rate 10/mnt/IP)
      operationId: googleLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '422':
          $ref: '#/components/responses/ValidationError'
        '200':
          description: Sukses; profileCompleted=false berarti FE wajib ke /onboarding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleLoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Request tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Token hilang / tidak valid / kedaluwarsa
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    ValidationError:
      description: Validasi request gagal (tipe/format field salah)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    RateLimited:
      description: Melebihi batas request per menit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    BadGateway:
      description: Layanan hulu (Google) gagal dijangkau
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ServiceUnavailable:
      description: Fitur belum dikonfigurasi di server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Conflict:
      description: Konflik status (mis. sudah di-sign / double request)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Token tidak valid atau kedaluwarsa.
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Role:
      type: string
      enum: [SUPER_ADMIN, ORG_ADMIN, SIGNER, VERIFIER]
...
    RegisterRequest:
      type: object
      required: [fullName, email, password, organization, role, purpose]
      properties:
        fullName:
          type: string
          minLength: 3
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        organization:
          type: string
          minLength: 2
          maxLength: 100
        phone:
          type: string
          description: Opsional saat register; format internasional, dinormalisasi
        role:
          type: string
          enum: [ORG_ADMIN, SIGNER]
        purpose:
          type: string
          minLength: 3
          description: Tujuan penggunaan; dicatat di audit log, bukan kolom user
        captchaToken:
          type: string
          description: Token Turnstile; wajib bila CAPTCHA diaktifkan
    VerifyEmailRequest:
      type: object
      required: [email, token]
      properties:
        email:
          type: string
          format: email
        token:
          type: string
    ResendVerificationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        captchaToken:
          type: string
          description: Token Turnstile; diverifikasi bila CAPTCHA diaktifkan
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        captchaToken:
          type: string
          description: Opsional; bila diisi akan diverifikasi
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    User:
      type: object
      required: [id, fullName, email, organization, role, emailVerified, authProvider, profileCompleted, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        organization:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        emailVerified:
          type: boolean
        phone:
          type: string
          nullable: true
        authProvider:
          type: string
          enum: [EMAIL, GOOGLE]
        profileCompleted:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Detik hingga access token kedaluwarsa (900 = 15 menit)
    LoginResponse:
      type: object
      required: [user, tokens]
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
    GoogleLoginRequest:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          description: Google ID Token dari GIS button
...
    GoogleLoginResponse:
      type: object
      required: [user, tokens, profileCompleted]
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
        profileCompleted:
          type: boolean
          description: False = FE wajib arahkan ke /onboarding
...
security:
  - bearerAuth: []